Fixed some buffer size stuff that stopped the program from working with more than 100 characters. Also chaged some data types too.

+++ Measure_Typing_Speed.c
@@ -6,7 +6,7 @@
 #include "Find_Faults.h"
 
 
-int Measure_Typing_Speed(char * Sentence,int SentenceLength,float * Words_Count)
+int Measure_Typing_Speed(char * Sentence,int SentenceLength,int* Words_Count)
 {
     fflush(stdout);
     char *UserSentence = (char *)malloc(SentenceLength*sizeof(char *));
