+++ Measure_Typing_Speed.h
@@ -1,1 +1,1 @@
-int Measure_Typing_Speed(char Sentence[],int SentenceLength,float * Words_Count);
+int Measure_Typing_Speed(char Sentence[],int SentenceLength,int* Words_Count);
